[data-client] move peer state and scoring into new file
